Use 64-bit selectors.
Since we don't free selectors, using 64-bit selectors is the better
option in any case. Additionally, it has performance advantages on
64-bit architectures.

+++ src/include/base/Types.h
@@ -39,7 +39,7 @@ typedef uint32_t label_t;
 #else
 typedef uint64_t label_t;
 #endif
-typedef unsigned capsel_t;
+typedef uint64_t capsel_t;
 typedef int fd_t;
 typedef uint64_t cycles_t;
